requirements.txt
pyhcl==0.3.13